% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_particle_size_df.R
\name{by_particle_size_df}
\alias{by_particle_size_df}
\title{By Particle Size Data Frame}
\usage{
by_particle_size_df(scenario)
}
\arguments{
\item{scenario}{list; A scenario definition list.}
}
\value{
a scenario data frame
}
\description{
Create a data frame of a scenario that varies by chute
particle size.
}
