% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_slope_df.R
\name{by_slope_df}
\alias{by_slope_df}
\title{By Slope Data Frame}
\usage{
by_slope_df(scenario)
}
\arguments{
\item{scenario}{list; A scenario definition list.}
}
\value{
a scenario data frame
}
\description{
Create a data frame of a scenario that varies by chute slope.
}
