% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_width_df.R
\name{by_width_df}
\alias{by_width_df}
\title{By Width Data Frame}
\usage{
by_width_df(scenario)
}
\arguments{
\item{scenario}{list; A scenario definition list.}
}
\value{
a scenario data frame
}
\description{
Create a data frame of a scenario that varies by chute width.
}
