% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_channel.R
\name{calc_channel}
\alias{calc_channel}
\title{Create a Channel Data Frame}
\usage{
calc_channel(channel_df)
}
\arguments{
\item{channel_df}{data.frame; A channel data frame where each row
represents a channel.}
}
\description{
Create a channel data frame.
}
