% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_dimensions.R
\name{channel_dimensions}
\alias{channel_dimensions}
\title{Create a Channel Data Frame}
\usage{
channel_dimensions(channel_df)
}
\arguments{
\item{channel_df}{data.frame; Input data frame with required channel variables}
}
\value{
A data frame with derived channel dimensions and stone parameters
}
\description{
Compute hydraulic and stone sizing metrics for each channel
}
