% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stone_size.R
\name{compute_stone_size}
\alias{compute_stone_size}
\title{Compute Stone Size}
\usage{
compute_stone_size(
  method,
  unit_discharge,
  slope,
  normal_velocity,
  stone_specific_weight,
  h2o_specific_weight,
  g = 9.81
)
}
\arguments{
\item{method}{character; The method to use for calculation. One of
"nrcs", "usace", "abt_johnson", "isbash", "usbr".}

\item{unit_discharge}{numeric; Unit discharge (e.g., m^2/s).}

\item{slope}{numeric; Channel slope (unitless).}

\item{normal_velocity}{numeric; Normal velocity (m/s).}

\item{stone_specific_weight}{numeric; Specific weight of stone (N/m^3).}

\item{h2o_specific_weight}{numeric; Specific weight of water (N/m^3).}

\item{g}{numeric; Acceleration due to gravity (default: 9.81 m/s^2).}
}
\value{
numeric; The calculated stone size (m).
}
\description{
Calculate stone sizes using specified methods.
}
\details{
The following methods provide approaches for estimating the stone size
required to resist motion under given hydraulic conditions. These methods
have been extensively documented in several publications:
\itemize{
\item \emph{Stone Sizing Criteria}, Natural Resources Conservation Service,
National Engineering Handbook 210 H Part 654 - Stream Restoration Design,
\href{https://directives.nrcs.usda.gov//sites/default/files2/1720623369/Technical\%20Supplement\%2014\%20C\%20-\%20Stone\%20Sizing\%20Criteria.pdf}{Technical Supplement 14 C},
August 2007.
\item \emph{Riprap Design Criteria, Recommended Specifications, and Quality Control},
National Cooperative Highway Research Program,
\href{https://onlinepubs.trb.org/onlinepubs/nchrp/nchrp_rpt_568.pdf}{NCHRP Report 568},
2006.
\item \emph{Hydraulic Design of Flood Control Channels}, US Army Corps of Engineers,
\href{https://www.publications.usace.army.mil/Portals/76/Publications/EngineerManuals/EM_1110-2-1601.pdf}{Engineering Manual 1110-2-1601},
30 June 1994.
}

\strong{Isbash Method Equation:}
\deqn{
  d = \frac{(1.2 \cdot v_c)^2}{2g \cdot (SG - 1)}
}
where:
\itemize{
\item \eqn{d}: Stone diameter (m)
\item \eqn{v_c}: Critical velocity for stone motion (m/s)
\item \eqn{SG}: Specific gravity = \eqn{(\frac{\text{stone\_specific\_weight}}{\text{h2o\_specific\_weight}})}
\item \eqn{g}: Gravitational acceleration (m/s²).
}
}
