% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_positron_extension.R
\name{install_positron_extension}
\alias{install_positron_extension}
\title{Install Positron Extension}
\usage{
install_positron_extension(extension_url)
}
\arguments{
\item{extension_url}{character; The URL to an extension VSIX file.}
}
\value{
Nothing. Installs the Positron extension.
}
\description{
Manually installs the requested Positron extension.
}
\details{
Check \href{https://open-vsx.org/?search=posit&sortBy=relevance&sortOrder=desc}{open-vsx.org}
for the latest versions of extensions available for Positron.
Find the \code{Download} button, copy the link, and provide that
url to this function.
}
