% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_isbash.R
\name{m_isbash}
\alias{m_isbash}
\title{Calculate Stone Size: Isbash Method}
\usage{
m_isbash(
  unit_discharge,
  slope,
  normal_velocity,
  stone_specific_weight,
  h2o_specific_weight,
  g = 9.81
)
}
\arguments{
\item{unit_discharge}{numeric; Unit discharge (e.g., m^2/s).}

\item{slope}{numeric; Channel slope (unitless).}

\item{normal_velocity}{numeric; Normal velocity (m/s).}

\item{stone_specific_weight}{numeric; Specific weight of stone (N/m^3).}

\item{h2o_specific_weight}{numeric; Specific weight of water (N/m^3).}

\item{g}{numeric; Acceleration due to gravity (default: 9.81 m/s^2).}
}
\value{
numeric; The calculated stone size (m).
}
\description{
Calculate stone size using the Isbash method.
}
\details{
\strong{Isbash Method}
The Isbash formula (Isbash	1936)	was	developed	for	the	construction of	dams
by	depositing rocks	into	moving	water. The Isbash curve	should only	be
used	for	quick	estimates	or	for	comparisons. A	coefficient	is	provided	to
target	high-	and	low-turbulence	flow	conditions,	so	this	method	can	be	a
high- or	low-energy	application.

\deqn{
  V_c = C \cdot ( 2 \cdot g \cdot \frac{\gamma_s - \gamma_w}{\gamma_w} ) ^{0.50} \cdot (D_{50})^{0.50}
}

\deqn{
  D_{50} = \frac{(C \cdot v_c)^2}{2g \cdot (SG - 1)}
}
where:
\itemize{
\item \eqn{D_{50}}: Median stone diameter (\eqn{ft})
\item \eqn{C}: Chezy coefficient. Smooth channels = 1.2, Rough channels = 0.86
\item \eqn{v_c}: Critical velocity for stone motion (\eqn{ft/s})
\item \eqn{\gamma_s}: stone density (\eqn{lb/ft^3})
\item \eqn{\gamma_w}: water density (\eqn{lb/ft^3})
\item \eqn{SG}: Specific gravity = \eqn{\frac{\gamma_s - \gamma_w}{\gamma_w}}
\item \eqn{g}: Acceleration due to gravity, (\eqn{ft/sec})
}

U.S. Department of Agriculture, Natural Resources Conservation Service. (2007).
\emph{Stone Sizing Criteria},
\href{https://directives.nrcs.usda.gov//sites/default/files2/1720623369/Technical\%20Supplement\%2014\%20C\%20-\%20Stone\%20Sizing\%20Criteria.pdf}{Technical Supplement 14 C},
National Engineering Handbook 210 Part 654 Stream Restoration Design, p.7
}
