% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_channel_flow.R
\name{plot_channel_flow}
\alias{plot_channel_flow}
\title{Plot Channel Flow Variables}
\usage{
plot_channel_flow(channel_dims, x_axis)
}
\arguments{
\item{channel_dims}{data frame; A channel dimensions data frame.}

\item{x_axis}{character; The variable to use for the x-axis. One of
"width", "slope", "particle_size"}
}
\value{
a ggplot object
}
\description{
Create a plot of channel flow variables from a channel
             dimensions data frame.
}
