% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stone_quantities.R
\name{plot_stone_quantities}
\alias{plot_stone_quantities}
\title{Plot Channel Stone Quantities}
\usage{
plot_stone_quantities(channel_dims, x_axis)
}
\arguments{
\item{channel_dims}{data frame; A channel dimensions data frame.}

\item{x_axis}{character; The variable to use for the x-axis. One of
"width", "length", "slope", "particle_size"}
}
\value{
a ggplot object
}
\description{
Create a plot of estimated stone quantities from a channel
dimensions data frame.
}
