% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stone_size_method.R
\name{plot_stone_size_method}
\alias{plot_stone_size_method}
\title{Plot Stone Size by Methods}
\usage{
plot_stone_size_method(stone_metrics, x_axis)
}
\arguments{
\item{stone_metrics}{data frame; A long-format data frame generated by 
`compute_stone_metrics`. Must include `method` 
and `stone_diameter` columns.}

\item{x_axis}{character; The variable to use for the x-axis. One of
"width", "length", "slope", "particle_size"}
}
\value{
a ggplot2 object
}
\description{
Plot estimated stone size according to several different
             methods.
}
