% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_dimensions.R
\name{channel_dimensions}
\alias{channel_dimensions}
\title{Create a Channel Data Frame}
\usage{
channel_dimensions(channel_df)
}
\arguments{
\item{channel_df}{data.frame; A channel data frame where each row
represents a channel.}
}
\value{
a data frame of channel dimensions
}
\description{
Create a channel data frame.
}
