% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stone_size.R
\name{compute_stone_size}
\alias{compute_stone_size}
\title{Compute Stone Size}
\usage{
compute_stone_size(
  method,
  unit_discharge,
  slope,
  normal_velocity,
  stone_specific_weight,
  h2o_specific_weight,
  g = 9.81
)
}
\arguments{
\item{method}{character; The method to use for calculation. One of
"nrcs", "usace", "abt_johnson", "isbash", "usbr".}

\item{unit_discharge}{numeric; Unit discharge (e.g., m^2/s).}

\item{slope}{numeric; Channel slope (unitless).}

\item{normal_velocity}{numeric; Normal velocity (m/s).}

\item{stone_specific_weight}{numeric; Specific weight of stone (N/m^3).}

\item{h2o_specific_weight}{numeric; Specific weight of water (N/m^3).}

\item{g}{numeric; Acceleration due to gravity (default: 9.81 m/s^2).}
}
\value{
numeric; The calculated stone size (m).
}
\description{
Calculate stone sizes using specified methods.
}
