% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scenario.R
\name{create_scenario}
\alias{create_scenario}
\title{Create Scenario}
\usage{
create_scenario(
  width,
  width_start,
  width_end,
  width_by,
  length,
  length_start,
  length_end,
  length_by,
  slope,
  slope_start,
  slope_end,
  slope_by,
  particle_size,
  particle_size_start,
  particle_size_end,
  particle_size_by,
  side_slope,
  total_discharge,
  stone_density,
  contingency,
  porosity,
  water_density,
  gravity
)
}
\arguments{
\item{width_start}{numeric; Width of the chute}

\item{gravity}{}
}
\description{
Create a scenario.
}
