% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stone_size_method.R
\name{plot_stone_size_method}
\alias{plot_stone_size_method}
\title{Plot Stone Size by Methods}
\usage{
plot_stone_size_method(channel_dims)
}
\arguments{
\item{channel_dims}{data frame; A channel dimensions data frame.}
}
\value{
a ggplot2 object
}
\description{
Plot estimated stone size according to several different
             methods.
}
